+++ projects/models.py
@@ -19,13 +19,17 @@ class RoleProjectManager(models.Manager):
         :param description: descripcion del rol
         :param permissions_list: lista de permisos a ser asignados al rol
         """
-        role = RoleProject.objects.create(
-            role_name=name,
-            description=description,
-            project=Project.objects.get(id=id_project)
-        )
-        self.attach_permissions(role.id, permissions_list)
-        print('rol creado exitosamente')
+        if not RoleProject.objects.filter(role_name=name).exists():
+            role = RoleProject.objects.create(
+                role_name=name,
+                description=description,
+                project=Project.objects.get(id=id_project)
+            )
+            self.attach_permissions(role.id, permissions_list)
+            print('rol creado exitosamente')
+            return True
+        else:
+            return False
 
     # funcion encargada de asignar permisos a los roles
     def attach_permissions(self, id_role, permissions_list):
@@ -56,7 +60,7 @@ class RoleProjectManager(models.Manager):
 
         :param id_project: id del proyecto
         """
-        return RoleProject.objects.filter(project_id=id_project)
+        return RoleProject.objects.filter(project_id=id_project).order_by('id')
 
     def get_member_roles(self, id_user, id_project):
         """
@@ -76,8 +80,13 @@ class RoleProjectManager(models.Manager):
         :param id_role: id del rol a ser eliminado
         """
         role = RoleProject.objects.get(id=id_role)
-        role.delete()
-        print("Eliminado exitosamente")
+        if not ProjectMember.objects.filter(
+                roles__id=id_role).count() > 0 and role.role_name != UProjectDefaultRoles.SCRUM_MASTER:
+            role.delete()
+            print("Eliminado exitosamente")
+            return True
+        else:
+            return False
 
     @staticmethod
     def update_role(id_role, perms, name=None, description=None):
@@ -177,7 +186,7 @@ class RoleProjectManager(models.Manager):
         """
         name = UProjectDefaultRoles.SCRUM_MASTER
         description = "Este rol es de Scrum Master"
-        permissions_list = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23]
+        permissions_list = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,24,25]
         rol = RoleProject.objects.create(role_name=name, description=description, project=None)
         rol.perms.set(permissions_list)
         return rol
@@ -232,8 +241,9 @@ class ProjectManager(models.Manager):
             user_id=user_id
         )
         member.roles.add(*roles)
+        return member
 
-    def delete_member(self, user_id,project):
+    def delete_member(self, user_id, project):
         """
 
         :param user_id:
@@ -253,6 +263,7 @@ class ProjectManager(models.Manager):
             project.members.remove(member.user)
             return True
         return False
+
     def get_project_members(self, id_project):
         return ProjectMember.objects.filter(project_id=id_project)
